+++ docs/examples/markdown/clear.md
@@ -1,0 +1,16 @@
+```python
+>>> x = 1
+>>> x
+1
+```
+
+Now let's start a new test:
+
+<!-- clear-namespace -->
+
+```python
+>>> x
+Traceback (most recent call last):
+...
+NameError: name 'x' is not defined
+```
