+++ docs/examples/markdown/codeblock-bash.md
@@ -1,0 +1,4 @@
+```bash
+$ echo hi there
+hi there
+```
