+++ docs/examples/markdown/codeblock-python.md
@@ -1,0 +1,13 @@
+<!-- invisible-code-block: python 
+
+# This could be some state setup needed to demonstrate things
+initialized = True
+-->
+
+This fenced code block defines a function:
+
+```python
+
+def prefix(text: str) -> str:
+    return 'prefix: '+text
+```
