+++ docs/examples/markdown/doctest.md
@@ -1,0 +1,7 @@
+A fenced code block:
+
+```python
+>>> x = 1+1
+>>> x
+2
+```
