+++ docs/index.rst
@@ -12,6 +12,7 @@ is designed so that it's easy to provide your own additional parsers.
    use.rst
    patterns.rst
    ReST Parsers <rest.rst>
+   markdown.rst
    myst.rst
 
 .. toctree::
