+++ sybil/parsers/markdown/skip.py
@@ -1,15 +1,13 @@
 from ..abstract import AbstractSkipParser
-from .lexers import DirectiveInPercentCommentLexer
 from ..markdown.lexers import DirectiveInHTMLCommentLexer
 
 
 class SkipParser(AbstractSkipParser):
     """
-    A :any:`Parser` for :ref:`skip <myst-skip-parser>` instructions.
+    A :any:`Parser` for :ref:`skip <markdown-skip-parser>` instructions.
     """
 
     def __init__(self) -> None:
         super().__init__([
-            DirectiveInPercentCommentLexer('skip'),
             DirectiveInHTMLCommentLexer('skip'),
         ])
