+++ tests/functional/markdown/doctest.md
@@ -1,0 +1,9 @@
+# DocTest examples
+
+A Python REPL / DocTest example using normal way of specifying Python
+language:
+
+```python
+>>> 1+1
+2
+```
